external/webkit/LayoutTests/platform/android/webgl-conformance/dumpAsText/conformance/glsl/implicit/ternary_ivec2_vec2.vert-expected.txt

CONSOLE MESSAGE: line 24: *** Error compiling shader '[object WebGLShader]':ERROR: 0:4: ':' :  wrong operand types  no operation ':' exists that takes a left-hand operand of type 'const mediump 2-component vector of float' and a right operand of type 'const mediump 2-component vector of int' (or there is no acceptable conversion)
ERROR: 0:4: '=' :  cannot convert from 'const mediump 2-component vector of int' to '2-component vector of float'

implicit cast of ivec2 to vec2 in ternary expression should fail

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

test: implicit cast of ivec2 to vec2 in ternary expression should fail
test vertex shader
default fragment shaderPASS implicit cast of ivec2 to vec2 in ternary expression should fail
PASS successfullyParsed is true

TEST COMPLETE
